+++ R/forest.multisimsum.R
@@ -81,6 +81,13 @@ forest.multisimsum <- function(obj, sstat, par = NULL, by = NULL, target = NULL,
     }
   }
 
+  ### Factorise `methodvar` if defined and if it is not already a factor to obtain a proper colour scale
+  if (!is.null(obj[["methodvar"]])) {
+  	if (!("factor" %in% class(obj[["summ"]][[obj[["methodvar"]]]]))) {
+  		obj[["summ"]][[obj[["methodvar"]]]] <- stats::relevel(factor(obj[["summ"]][[obj[["methodvar"]]]]), ref = obj[["ref"]])
+  	}
+  }
+
   ### Build a ggplot object
   if (is.null(par)) {
     gg <- ggplot2::ggplot(get_data(summary(obj))[get_data(obj)[["stat"]] == sstat, ], ggplot2::aes_string(x = obj[["methodvar"]], y = "est"))