+++ NEWS.md
@@ -1,5 +1,8 @@
 # rsimsum 0.3.0-9000
 
+Bug fixes:
+* `bar()`, `forest()`, `lolly()`, `heat()` now appropriately pick a discrete X (or Y) axis scale for methods (if defined) when the method variable is numeric.
+
 # rsimsum 0.3.0
 
 Bug fixes: